pyproject.toml
[tool.poetry]
name = "ddi-datenbankdesign-und-implementierung"
version = "0.1.0"
description = ""
authors = ["BR4GER <benjamin.wuermli@students.fhnw.ch>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.36"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
